refactoring table components to make them more readable
