+++ src/components/Planning/PlanningClassesTable/NameTooltip/NameTooltip.tsx
@@ -1,0 +1,17 @@
+import { FC, memo } from 'react';
+import './styles.css';
+
+interface INameTooltipProps {
+  value: string;
+}
+
+const NameTooltip: FC<INameTooltipProps> = ({ value }) => {
+  return (
+    <section className="tooltip-container">
+      {value}
+      <div className="tooltip-arrow" />
+    </section>
+  );
+};
+
+export default memo(NameTooltip);
