+++ src/components/Planning/PlanningClassesTable/NameTooltip/index.ts
@@ -1,0 +1,1 @@
+export { default as NameTooltip } from './NameTooltip';
