+++ src/components/Planning/PlanningClassesTable/PlanningClassesTableRow/index.ts
@@ -1,0 +1,1 @@
+export { default as PlanningClassesTableRow } from './PlanningClassesTableRow';
