+++ src/components/Planning/PlanningClassesTable/styles.css
@@ -73,118 +73,6 @@
   @apply font-bold text-suomenlinna;
 }
 
-.table-cell .table-cell-container span:nth-child(3) {
-  font-weight: bold;
-  color: var(--color-suomenlinna-dark);
-}
-
-.table-cell span {
-  font-weight: 300;
-  transform: translateX(-15%);
-}
-
-.table-cell.class.first {
-  background: var(--color-bus);
-}
-
-.table-cell.class.second {
-  background: rgb(0, 0, 163);
-}
-
-.table-cell.class.third {
-  background: rgb(0, 0, 129);
-}
-
-.table-cell.location.second {
-  background: var(--color-fog-medium-light);
-}
-
-.table-cell.location.third {
-  background: var(--color-coat-of-arms-medium-light);
-}
-
-.table-cell.project {
-  background: rgb(0, 0, 102);
-}
-
-.table-header-cell .table-header-content {
-  display: flex;
-  align-items: center;
-  text-overflow: ellipsis;
-}
-
-.table-header-content.first {
-  padding-left: 0;
-}
-
-.table-header-content.second {
-  padding-left: 2rem;
-}
-
-.table-header-content.third {
-  padding-left: 4rem;
-}
-
-.table-header-content.project {
-  padding-left: 6rem;
-}
-
-.table-header-content-dots {
-  display: flex;
-  margin-right: 0.5rem;
-}
-
-.table-header-content-item {
-  display: flex;
-  margin-right: 0.5rem;
-  position: relative;
-}
-
-.table-header-content-item:hover {
-  cursor: pointer;
-}
-
-.custom-tooltip button {
-  width: 14rem !important;
-  position: absolute;
-  left: 12rem;
-}
-
-.custom-tooltip section {
-  color: black !important;
-}
-
-.custom-tooltip svg {
-  color: transparent;
-}
-
-.tooltip-container {
-  inset: auto auto 0px 0px;
-  @apply absolute z-[1000] hidden max-w-[21.875rem] -translate-y-8 
-  border-b-8 border-coat-of-arms bg-coat-of-arms-l p-4 text-sm font-normal leading-normal text-black;
-}
-
-.tooltip-arrow {
-  --arrow-pos: calc(-1 * var(--spacing-s) - 8px);
-  @apply absolute bottom-[var(--arrow-pos)] mx-auto hidden border-8 border-transparent border-t-coat-of-arms;
-}
-
-.table-title-container {
-  max-width: 15rem;
-}
-
-/* hide too long class names */
-.table-title-container span {
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  overflow: hidden;
-  display: block;
-}
-
-.table-header-content-item:hover .tooltip-container {
-  display: block;
-}
-
-.table-header-content-item:hover .tooltip-arrow {
-  display: block;
+.planning-table thead tr th:nth-child(2) span {
+  padding-right: 0;
 }
